Graph implemented showing Target value, end timescale and total investment

+++ Code/InvestmentProjection/InvestmentProjection.BusinessLogic/Models/GraphDataSet.cs
@@ -2,7 +2,7 @@
 {
     public class GraphDataSet
     {
-        public int X { get; set; }
+        public decimal X { get; set; }
 
         public decimal Y { get; set; }
     }
